Complete codegen of 'build' workflow YAML tree (#38631)
Summary:
This replaces all "verbatim-sources" files comprising the workflow named 'build' in the CircleCI config with code generation.  This shall facilitate an automated conversion to workflow-per-job.

Note that the '.circleci/config.yml' file has some strictly cosmetic changes in this PR: some keys are sorted and inline comments are removed (moved to the Python modules that generate the config).
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38631

Differential Revision: D21623528

Pulled By: kostmo

fbshipit-source-id: d86bd7aea979f443db14b4a3898220faad6bd0da