[vulkan] VulkanTensor, add strides in interface (#39077)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39077

We plan to support strides for Vulkan, but that is not implemented yet.
The main intention of this faking of strides and is_contiguous - to be able to run torchscript mobilenetV2 on Vulkan backend for development and profiling.

This change adds strides to Vulkan interface and overrides strides(), stride(), is_contiguous() of OpaqueTensorImpl for that purpose.

Test Plan: Imported from OSS

Differential Revision: D21962426

Pulled By: IvanKobzarev

fbshipit-source-id: cfef4903ad7062170926264f45cff1293ade78f6